requirements.txt
pandas-2.2.3
seaborn-0.13.2
matplotlib-3.10.3
xlrd-2.0.1
openpyxl-3.1.5
PyMySQL-1.1.1
sqlalchemy-2.0.41
jinja2-3.1.6
rapidfuzz-3.13.0
scipy-1.15.3
patsy-1.0.1
statsmodels-0.14.4
altair-5.5.0
attrs-25.3.0
blinker-1.9.0
cachetools-5.5.2
certifi-2025.4.26
charset-normalizer-3.4.2
click-8.2.1
gitdb-4.0.12
gitpython-3.1.44
idna-3.10
jsonschema-4.24.0
jsonschema-specifications-2025.4.1
narwhals-1.41.0
packaging-24.2
protobuf-6.31.1
pyarrow-20.0.0
pydeck-0.9.1
referencing-0.36.2
requests-2.32.3
rpds-py-0.25.1
smmap-5.0.2
streamlit-1.45.1
tenacity-9.1.2
toml-0.10.2
urllib3-2.4.0
branca-0.8.1
folium-0.19.6
xyzservices-2025.4.0
streamlit_folium-0.25.0
shapely-2.1.1